requirements.txt
google-api-python-client==2.109.0
google-auth==2.23.4
google-auth-oauthlib==1.0.0
google-auth-httplib2==0.1.1
python-dotenv==1.0.0
click==8.1.7
google-genai==0.3.0
pydantic==2.5.0
httpx==0.25.0
structlog==23.2.0